+++ previous-versions/commerce/src/codatcommerce/sdkconfiguration.py
@@ -20,9 +20,9 @@ class SDKConfiguration:
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = '3.0.0'
-    sdk_version: str = '0.34.0'
-    gen_version: str = '2.150.0'
-    user_agent: str = 'speakeasy-sdk/python 0.34.0 2.150.0 3.0.0 codat-commerce'
+    sdk_version: str = '0.34.1'
+    gen_version: str = '2.155.1'
+    user_agent: str = 'speakeasy-sdk/python 0.34.1 2.155.1 3.0.0 codat-commerce'
     retry_config: RetryConfig = None
 
     def get_server_details(self) -> tuple[str, dict[str, str]]: