+++ previous-versions/commerce/RELEASES.md
@@ -627,3 +627,13 @@ Based on:
 - [python v0.34.0] previous-versions/commerce
 ### Releases
 - [PyPI v0.34.0] https://pypi.org/project/codat-commerce/0.34.0 - previous-versions/commerce
+
+## 2023-10-16 01:27:59
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Commerce.yaml
+- Speakeasy CLI 1.99.2 (2.155.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.34.1] previous-versions/commerce
+### Releases
+- [PyPI v0.34.1] https://pypi.org/project/codat-commerce/0.34.1 - previous-versions/commerce